app/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
black==19.3b0
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dash==0.37.0
dash-core-components==0.43.1
dash-html-components==0.13.5
dash-renderer==0.18.0
dash-table==3.4.0
dash-table-experiments==0.6.0
decorator==4.4.0
Flask==1.0.2
Flask-Compress==1.4.0
Flask-Cors==3.0.6
gunicorn==19.7.1
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
kaggle==1.5.3
lxml==4.3.3
MarkupSafe==1.1.1
more-itertools==7.0.0
nbformat==4.4.0
numpy==1.16.2
pandas==0.23.4
pandas-datareader==0.3.0.post0
Pillow==5.2.0
plotly==3.6.1
pluggy==0.6.0
py==1.8.0
pycosat==0.6.3
pycparser==2.19
pydicom==1.2.2
pyOpenSSL==19.0.0
pyrsistent==0.14.11
PySocks==1.6.8
pytest==3.6.0
python-dateutil==2.8.0
python-dotenv==0.8.2
python-slugify==3.0.2
pytz==2019.1
requests==2.21.0
requests-file==1.4.3
requests-ftp==0.3.1
retrying==1.3.3
ruamel-yaml==0.15.46
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
text-unidecode==1.2
toml==0.10.0
tqdm==4.31.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.24.2
Werkzeug==0.15.2
